+++ airbyte-webapp/src/components/ui/Markdown/Markdown.module.scss
@@ -1,5 +1,5 @@
-@use "../../scss/colors";
-@use "../../scss/fonts";
+@use "../../../scss/colors";
+@use "../../../scss/fonts";
 
 .markdown {
   * {
