+++ airbyte-webapp/src/components/ui/Markdown/admonitions.module.scss
@@ -1,4 +1,4 @@
-@use "../../scss/colors";
+@use "../../../scss/colors";
 
 .admonition {
   border-radius: 8px;
